fix: use correct types for Windows MSVC FFI callbacks
- Change on_stream_read_callback and on_select_padding return types from
  isize to i64 to match nghttp2 expected callback signatures on Windows
- Cast sock_addr.len() to u32 in uv_tcp_bind calls (socket2 returns i32
  on Windows)

+++ ext/node/ops/libuv_stream.rs
@@ -392,7 +392,7 @@ impl TCP {
       uv_compat::uv_tcp_bind(
         tcp,
         sock_addr.as_ptr() as *const _,
-        sock_addr.len(),
+        sock_addr.len() as u32,
         0,
       )
     }
@@ -419,7 +419,7 @@ impl TCP {
       uv_compat::uv_tcp_bind(
         tcp,
         sock_addr.as_ptr() as *const _,
-        sock_addr.len(),
+        sock_addr.len() as u32,
         0,
       )
     }
